new: added VisualParity validation state and reorganized validations
